requirements.txt
openai==0.27.8
python-dotenv==1.0.0
python-docx==0.8.11
matplotlib==3.7.3
python-bidi==0.4.2
python-telegram-bot==20.3